Issue #29: Rename Scruby to SRuby
