--- src/main/scala/com/ethowitz/scruby/exceptions/CompilationError.scala
@@ -1,3 +1,0 @@
-package com.ethowitz.scruby.exceptions
-
-trait CompilationError
