+++ src/main/scala/com/ethowitz/sruby/core/RubyNilClass.scala
@@ -1,9 +1,9 @@
-package com.ethowitz.scruby.core
+package com.ethowitz.sruby.core
 
-import com.ethowitz.scruby.evaluator.MethodMap
-import com.ethowitz.scruby.evaluator.RubyMethod
-import com.ethowitz.scruby.evaluator.VariableMap
-import com.ethowitz.scruby.parser.RubyObjectContainerNode
+import com.ethowitz.sruby.evaluator.MethodMap
+import com.ethowitz.sruby.evaluator.RubyMethod
+import com.ethowitz.sruby.evaluator.VariableMap
+import com.ethowitz.sruby.parser.RubyObjectContainerNode
 
 object RubyNilClass extends RubyObject('NilClass, None, MethodMap.empty, VariableMap.empty) {
   override def toString: String = "nil"
