+++ src/main/scala/com/ethowitz/sruby/core/RubyTrueClass.scala
@@ -1,7 +1,7 @@
-package com.ethowitz.scruby.core
+package com.ethowitz.sruby.core
 
-import com.ethowitz.scruby.evaluator.MethodMap
-import com.ethowitz.scruby.evaluator.VariableMap
+import com.ethowitz.sruby.evaluator.MethodMap
+import com.ethowitz.sruby.evaluator.VariableMap
 
 object RubyTrueClass extends RubyObject('TrueClass, None, MethodMap.empty, VariableMap.empty) {
   override def toString: String = "true"
