+++ src/main/scala/com/ethowitz/sruby/exceptions/CompilationError.scala
@@ -1,0 +1,3 @@
+package com.ethowitz.sruby.exceptions
+
+trait CompilationError
