+++ src/main/scala/com/ethowitz/sruby/exceptions/LexerError.scala
@@ -1,3 +1,3 @@
-package com.ethowitz.scruby.exceptions
+package com.ethowitz.sruby.exceptions
 
 final case class LexerError(msg: String) extends CompilationError
