+++ src/main/scala/com/ethowitz/sruby/exceptions/ParserError.scala
@@ -1,3 +1,3 @@
-package com.ethowitz.scruby.exceptions
+package com.ethowitz.sruby.exceptions
 
 final case class ParserError(msg: String) extends CompilationError
