+++ src/main/scala/com/ethowitz/sruby/lexer/Token.scala
@@ -1,4 +1,4 @@
-package com.ethowitz.scruby.lexer
+package com.ethowitz.sruby.lexer
 
 sealed trait Token
